split reduce_kernel
