+++ paddle/phi/kernels/reduce_mean_kernel.h
@@ -15,10 +15,12 @@
 #pragma once
 
 #include "paddle/phi/core/dense_tensor.h"
+#include "paddle/phi/infermeta/unary.h"
 
 namespace phi {
+
 template <typename T, typename Context>
-void ProdRawKernel(const Context& dev_ctx,
+void MeanRawKernel(const Context& dev_ctx,
                    const DenseTensor& x,
                    const std::vector<int64_t>& dims,
                    bool keep_dim,
@@ -27,11 +29,22 @@ void ProdRawKernel(const Context& dev_ctx,
 
 
 template <typename T, typename Context>
-void ProdKernel(const Context& dev_ctx,
+void MeanKernel(const Context& dev_ctx,
                 const DenseTensor& x,
                 const std::vector<int64_t>& dims,
                 bool keep_dim,
                 DenseTensor* out);
 
+template <typename T, typename Context>
+DenseTensor Mean(const Context& dev_ctx,
+                 const DenseTensor& x,
+                 const std::vector<int64_t>& axis,
+                 bool keep_dim) {
+  DenseTensor dense_out;
+  MetaTensor meta_out(&dense_out);
+  SumRawInferMeta(x, axis, keep_dim, false, x.dtype(), &meta_out);
+  MeanKernel<T, Context>(dev_ctx, x, axis, keep_dim, &dense_out);
+  return dense_out;
+}
 
 }  // namespace phi
